fix: add login and signup pages Figured it'd be better to have dedicated pages for users to login and signup to the system. This way, we can have a more user-friendly experience.
fix: move the components folder to the root directory
This way, we can have a more organized structure for our components. The components folder is now located in the root directory, making it easier to find and manage our components. Also, this is the standard practice for Next.js projects, so it aligns with the community conventions.

fix: create custom tailwind utility classes
This was a crazy one. I spent almost 6 hours trying to figure out how to create custom tailwind utility classes in tailwindcss v4. I kept getting an error not knowing that things work a bit differently in v4. I finally figured it out and now we have custom utility classes for our project. This will make it easier to maintain and update our styles in the future.